trigger_dataprep/requirements.txt
requests
google-cloud-storage==2.14.0